use pytest fixtures, docstring and typing adjustments, move yml files to data folders, add pandas to requirements
