+++ environment.yml
@@ -4,6 +4,7 @@ channels:
 dependencies:
   - python >=3.9,<3.12
   - numpy
+  - pandas
   - pyyaml
   - scipy
   - statsmodels
