+++ tests/test_hydrotel.py
@@ -6,62 +6,12 @@ import numpy as np
 import pandas as pd
 import pytest
 import xarray as xr
-from xclim.testing.helpers import test_timeseries as timeseries
 
 import xhydro
 from xhydro.modelling import Hydrotel
 
 
 class TestHydrotel:
-    @staticmethod
-    def make_meteo():
-        meteo = timeseries(
-            np.zeros(365 * 2),
-            start="2001-01-01",
-            freq="D",
-            variable="tasmin",
-            as_dataset=True,
-            units="degC",
-        )
-        meteo["tasmax"] = timeseries(
-            np.ones(365 * 2),
-            start="2001-01-01",
-            freq="D",
-            variable="tasmax",
-            units="degC",
-        )
-        meteo["pr"] = timeseries(
-            np.ones(365 * 2) * 10,
-            start="2001-01-01",
-            freq="D",
-            variable="pr",
-            units="mm",
-        )
-        meteo = meteo.expand_dims("stations").assign_coords(stations=["010101"])
-        meteo = meteo.assign_coords(coords={"lat": 46, "lon": -77, "z": 0})
-        for c in ["lat", "lon", "z"]:
-            meteo[c] = meteo[c].expand_dims("stations")
-        return meteo
-
-    @staticmethod
-    def make_debit_aval():
-        debit_aval = timeseries(
-            np.zeros(365 * 2),
-            start="2001-01-01",
-            freq="D",
-            variable="streamflow",
-            as_dataset=True,
-        )
-        debit_aval = debit_aval.expand_dims("troncon").assign_coords(troncon=[0])
-        debit_aval = debit_aval.assign_coords(coords={"idtroncon": 0})
-        debit_aval["idtroncon"] = debit_aval["idtroncon"].expand_dims("troncon")
-        debit_aval = debit_aval.rename({"streamflow": "debit_aval"})
-        debit_aval["debit_aval"].attrs = {
-            "units": "m3/s",
-            "description": "Debit en aval du troncon",
-        }
-        return debit_aval
-
     def test_options(self, tmpdir):
         xhydro.testing.utils.fake_hydrotel_project(tmpdir, "fake")
         ht = Hydrotel(
@@ -130,9 +80,9 @@ class TestHydrotel:
         assert ht2.output_options == outopt
 
     @pytest.mark.parametrize("test", ["station", "grid", "none", "toomany"])
-    def test_get_data(self, tmpdir, test):
-        meteo = self.make_meteo()
-        debit_aval = self.make_debit_aval()
+    def test_get_data(self, tmpdir, test, make_meteo, make_debit_aval):
+        meteo = make_meteo()
+        debit_aval = make_debit_aval()
 
         xhydro.testing.utils.fake_hydrotel_project(
             tmpdir, "fake", meteo=meteo, debit_aval=debit_aval
@@ -178,8 +128,8 @@ class TestHydrotel:
                 ht.get_input()
 
     @pytest.mark.parametrize("test", ["ok", "option", "file", "health"])
-    def test_basic(self, tmpdir, test):
-        meteo = self.make_meteo()
+    def test_basic(self, tmpdir, test, make_meteo):
+        meteo = make_meteo()
         if test == "health":
             meteo = meteo.reset_coords("z", drop=True).squeeze()
             meteo = meteo.convert_calendar("noleap")
@@ -245,8 +195,8 @@ class TestHydrotel:
             ):
                 ht._basic_checks()
 
-    def test_standard(self, tmpdir):
-        debit_aval = self.make_debit_aval()
+    def test_standard(self, tmpdir, make_debit_aval):
+        debit_aval = make_debit_aval()
         xhydro.testing.utils.fake_hydrotel_project(
             tmpdir, "fake", debit_aval=debit_aval
         )
@@ -321,10 +271,8 @@ class TestHydrotel:
         assert ht.simulation_options["ECRITURE ETAT FONTE NEIGE"] == "2001-01-01 00"
 
     @pytest.mark.parametrize("test", ["ok", "pdt", "cfg"])
-    def test_run(self, tmpdir, test):
-        xhydro.testing.utils.fake_hydrotel_project(
-            tmpdir, "fake", meteo=self.make_meteo()
-        )
+    def test_run(self, tmpdir, test, make_meteo):
+        xhydro.testing.utils.fake_hydrotel_project(tmpdir, "fake", meteo=make_meteo())
         ht = Hydrotel(
             tmpdir / "fake",
             default_options=False,
